requirements.txt
allure-pytest==2.11.1
Appium-Python-Client==2.7.1
pendulum==2.1.2
pytest==7.1.3
PyYAML==6.0
